Made the 0.6 version into the default create command. Added new option for a custom boilerplate.
