+++ lib/commands/create-0.5.js
@@ -1,6 +1,5 @@
 require('shelljs/global');
 var replace = require('replace');
-var prompt = require('prompt');
 var util = require('../util');
 var config = require('../../config');
 
@@ -8,14 +7,14 @@ module.exports = function (program) {
   var self = this;
 
   program
-    .command('create-0.6 <projectTitle>')
-    .description('Create a boilerplate Apostrophe 0.6 project')
-    .option('--setup', 'Will setup the project further by installing dependencies and creating an admin user')
+    .command('create-0.5 <projectTitle>')
+    .description('Create a boilerplate Apostrophe 0.5 project')
+    .option('--install', 'Will install dependencies after creating boilerplate')
     .action(function(projectTitle, options) {
-      var count = (options.install || options.setup) ? 5 : 2;
+      var count = (options.install) ? 5 : 2;
       util.nlog('create', 'Grabbing the boilerplate from Github [1/'+count+']' );
       // clone the sandbox project
-      if (exec('git clone ' + config.APOSTROPHE_BOILERPLATE_UNSTABLE + ' ' + projectTitle).code !== 0) {
+      if (exec('git clone ' + config.APOSTROPHE_BOILERPLATE + ' ' + projectTitle).code !== 0) {
         util.error('create');
         return false;
       }
@@ -31,7 +30,7 @@ module.exports = function (program) {
       replaceInConfig("https:\\/\\/github.com\\/punkave\\/apostrophe-boilerplate", "");
 
       // if we catch an install flag, do some stuff
-      if (options.install || options.setup){
+      if (options.install){
         util.nlog('create', 'Creating your local data file [3/'+count+']');
         mkdir('-p', 'data/');
         cp('local.example.js', 'data/local.js');
@@ -39,28 +38,13 @@ module.exports = function (program) {
         replaceInData(/[Aa]postrophe[\s\-][Ss]andbox/g, projectTitle);
         util.nlog('create', 'Installing all dependencies [4/'+count+']');
         exec('npm install');
-        // Create an admin user (note this will prompt for password)
-        util.nlog('create', 'Creating an admin user [5/5]');
-        util.nlog('create', 'Choose a password for the admin user');
-        prompt.start();
-        prompt.get({
-          properties: {
-            password: {
-              required: true,
-              hidden: true
-            }
-          }
-        }, function(err, result) {
-          if (err) {
-            util.error('create');
-            return false;
-          }
-          exec('echo "' + result.password + '" | node app.js apostrophe-users:add admin admin');
-          util.nlog('create', 'Login as "admin"');
-          util.success('create');
-          return true;
-        });
+        //Can I just call this from another module?
+        util.nlog('create', 'Resetting your database [5/'+count+']');
+        exec('node app apostrophe:reset');
       }
+
+      util.success('create');
+      return true;
     });
 }
