+++ lib/commands/create.js
@@ -1,5 +1,6 @@
 require('shelljs/global');
 var replace = require('replace');
+var prompt = require('prompt');
 var util = require('../util');
 var config = require('../../config');
 
@@ -8,13 +9,16 @@ module.exports = function (program) {
 
   program
     .command('create <projectTitle>')
-    .description('Create a boilerplate apostrophe project')
-    .option('--install', 'Will install dependencies after creating boilerplate')
+    .description('Create a boilerplate Apostrophe 0.6 project')
+    .option('--setup', 'Will setup the project further by installing dependencies and creating an admin user')
+    .option('--boilerplate [url]', 'Use a custom boilerplate to create the project with')
     .action(function(projectTitle, options) {
-      var count = (options.install) ? 5 : 2;
+      var count = (options.install || options.setup) ? 5 : 2;
+      var boilerplateUrl = (options.boilerplate) ? options.boilerplate : config.APOSTROPHE_BOILERPLATE_UNSTABLE;
+
       util.nlog('create', 'Grabbing the boilerplate from Github [1/'+count+']' );
       // clone the sandbox project
-      if (exec('git clone ' + config.APOSTROPHE_BOILERPLATE + ' ' + projectTitle).code !== 0) {
+      if (exec('git clone ' + boilerplateUrl + ' ' + projectTitle).code !== 0) {
         util.error('create');
         return false;
       }
@@ -30,7 +34,7 @@ module.exports = function (program) {
       replaceInConfig("https:\\/\\/github.com\\/punkave\\/apostrophe-boilerplate", "");
 
       // if we catch an install flag, do some stuff
-      if (options.install){
+      if (options.install || options.setup){
         util.nlog('create', 'Creating your local data file [3/'+count+']');
         mkdir('-p', 'data/');
         cp('local.example.js', 'data/local.js');
@@ -38,13 +42,28 @@ module.exports = function (program) {
         replaceInData(/[Aa]postrophe[\s\-][Ss]andbox/g, projectTitle);
         util.nlog('create', 'Installing all dependencies [4/'+count+']');
         exec('npm install');
-        //Can I just call this from another module?
-        util.nlog('create', 'Resetting your database [5/'+count+']');
-        exec('node app apostrophe:reset');
+        // Create an admin user (note this will prompt for password)
+        util.nlog('create', 'Creating an admin user [5/5]');
+        util.nlog('create', 'Choose a password for the admin user');
+        prompt.start();
+        prompt.get({
+          properties: {
+            password: {
+              required: true,
+              hidden: true
+            }
+          }
+        }, function(err, result) {
+          if (err) {
+            util.error('create');
+            return false;
+          }
+          exec('echo "' + result.password + '" | node app.js apostrophe-users:add admin admin');
+          util.nlog('create', 'Login as "admin"');
+          util.success('create');
+          return true;
+        });
       }
-
-      util.success('create');
-      return true;
     });
 }
